requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# HTTP & APIs
httpx==0.25.2
requests==2.31.0

# Environment & Config
python-dotenv==1.0.0

# Notion Integration
notion-client==2.0.0

# Gemini API
google-generativeai==0.3.0

# Async Support
anyio==3.7.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Retry logic
tenacity==8.2.3
